LRCI-7820 Extract BaseMonitor from the job health monitor
