+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/monitor/BaseMonitor.java
@@ -1,0 +1,101 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.jenkins.results.parser.monitor;
+
+import com.liferay.jenkins.results.parser.JenkinsResultsParserUtil;
+
+import java.util.Map;
+
+/**
+ * @author Brittney Nguyen
+ */
+public abstract class BaseMonitor implements Monitor {
+
+	@Override
+	public String getId() {
+		return _monitorConfig.getId();
+	}
+
+	@Override
+	public MonitorConfig getMonitorConfig() {
+		return _monitorConfig;
+	}
+
+	protected BaseMonitor(MonitorConfig monitorConfig) {
+		_monitorConfig = monitorConfig;
+	}
+
+	protected int getAttemptTimeoutMillis() {
+		long timeoutSeconds = _monitorConfig.getTimeoutSeconds();
+
+		if (timeoutSeconds <= 0) {
+			timeoutSeconds = _SECONDS_TIMEOUT_DEFAULT;
+		}
+
+		timeoutSeconds = Math.min(timeoutSeconds, Integer.MAX_VALUE / 1000);
+
+		return (int)((timeoutSeconds * 1000) / 3);
+	}
+
+	protected String getInvalidValueMessage(
+		String category, String name, String value) {
+
+		return JenkinsResultsParserUtil.combine(
+			"Invalid ", name, " for ", getKey(category, name), ": ", value);
+	}
+
+	protected String getKey(String category, String name) {
+		return JenkinsResultsParserUtil.combine(
+			"monitor[", _monitorConfig.getId(), "].", category, "[", name, "]");
+	}
+
+	protected long getLongValue(
+		String category, long defaultValue, String name,
+		Map<String, String> values) {
+
+		String value = values.get(name);
+
+		if (JenkinsResultsParserUtil.isNullOrEmpty(value)) {
+			return defaultValue;
+		}
+
+		long longValue = 0;
+
+		try {
+			longValue = Long.parseLong(value);
+		}
+		catch (NumberFormatException numberFormatException) {
+			throw new IllegalArgumentException(
+				getInvalidValueMessage(category, name, value),
+				numberFormatException);
+		}
+
+		if (longValue < 0) {
+			throw new IllegalArgumentException(
+				getInvalidValueMessage(category, name, value));
+		}
+
+		return longValue;
+	}
+
+	protected String getRequiredParameter(
+		String name, Map<String, String> parameters) {
+
+		String value = parameters.get(name);
+
+		if (JenkinsResultsParserUtil.isNullOrEmpty(value)) {
+			throw new IllegalArgumentException(
+				"Missing required property " + getKey("parameter", name));
+		}
+
+		return value;
+	}
+
+	private static final long _SECONDS_TIMEOUT_DEFAULT = 60;
+
+	private final MonitorConfig _monitorConfig;
+
+}
